Fix buffering problem
Previously, the message handler wouldn't process all of the messages
that we received over the TCP socket. It would read from the socket
via a FILE* and process the first message from that FILE*. Then it
would check (via poll) whether there was more data to read on the
socket file descriptor. However, the initial read from the socket
would often buffer data in userspace that would not be read. Since
it had already been read out of the file descriptor, the program
logic assumed that all data had been processed.

The solution for this (at least for the time being) is to make the
TCP socket nonblocking and to ensure that we do not stop reading
messages from the FILE* until it is no longer possible to read
characters from the FILE*.